+++ .gitignore
@@ -1,3 +1,7 @@
 __pycache__/*
-.devcontainer/devcontainer.json
 app/__pycache__/*
+app/api/__pycache__/*
+app/api/v1/__pycache__/*
+app/service/__pycache__/*
+app/models/__pycache__/*
+
