+++ app/models/models.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, Time, DateTime
 from sqlalchemy.orm import relationship
 
-from .database import Base
+from ..service.database import Base
 
 class User(Base):
     __tablename__ = "users"
