+++ app/service/controller.py
@@ -7,23 +7,24 @@ from passlib.context import CryptContext
 from sqlalchemy.orm import Session
 from jose import JWTError, jwt
 
-from . import crud, models, schemas
+from ..models import models, schemas
+
+from ..service import database_connection
 
 SECRET_KEY = "4757a5228c03cece3cfd8ce77e05e1769597b94bf416d934120bb0720c91f3ba"
 ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 60
-REFRESH_TOKEN_EXPIRE_MINUTES = 360
+
 
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/login")
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="login")
 
 # Dependency
 def get_db(request: Request):
     return request.state.db
 
 def get_user(username: str, db: Session = Depends(get_db)):
-    db_user = crud.get_user_by_username(db, username=username)
+    db_user = database_connection.get_user_by_username(db, username=username)
     if db_user is None:
         raise HTTPException(status_code=404, detail="User not found")
     return db_user
@@ -35,7 +36,7 @@ def get_password_hash(password):
     return pwd_context.hash(password)
 
 def authenticate_user(db, username: str, password: str):
-    user = crud.get_user_by_username(db, username=username)
+    user = database_connection.get_user_by_username(db, username=username)
     if not user:
         return False
     if not verify_password(password, user.hashed_password):
@@ -67,31 +68,10 @@ async def get_current_user(token: str = Depends(oauth2_scheme), db: Session = De
     except JWTError as e:
         print (f"An error occurred due to: '{e}'")
         raise credentials_exception
-    user = crud.get_user_by_username(db, username=token_data.username)
+    user = database_connection.get_user_by_username(db, username=token_data.username)
     if user is None:
         raise credentials_exception
     return user
-
-async def validate_refresh_token(token):
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("User")
-        if username is None:
-            raise credentials_exception
-        token_data = schemas.TokenData(username=username)
-    except JWTError as e:
-        print (f"An error occurred due to: '{e}'")
-        raise credentials_exception
-    user = crud.get_user_by_username(db, username=token_data.username)
-    if user is None:
-        raise credentials_exception
-    return user
-
 
 async def get_current_active_user(current_user: schemas.User = Depends(get_current_user)):
     if not current_user.is_active:
